Update backend: improvements and new features

+++ src/main/java/com/example/conges/repository/JoursPrisParTypeProjection.java
@@ -1,6 +1,7 @@
 package com.example.conges.repository;
 
 import com.example.conges.entity.TypeConge;
+import java.math.BigDecimal;
 
 /**
  * Projection pour la somme des {@code nombreJours} groupée par {@link TypeConge}.
@@ -9,5 +10,5 @@ public interface JoursPrisParTypeProjection {
 
     TypeConge getTypeConge();
 
-    Long getTotalJours();
+    BigDecimal getTotalJours();
 }
